Fix CTM and disable emissive for now
