+++ src/main/java/me/pepperbell/continuity/client/model/ModelObjectsContainer.java
@@ -7,9 +7,7 @@ import net.fabricmc.fabric.api.renderer.v1.mesh.MutableMesh;
 public class ModelObjectsContainer {
 	public static final ThreadLocal<ModelObjectsContainer> THREAD_LOCAL = ThreadLocal.withInitial(ModelObjectsContainer::new);
 
-	public final CtmBakedModel.CtmQuadTransform ctmQuadTransform = new CtmBakedModel.CtmQuadTransform();
-	public final EmissiveBakedModel.EmissiveBlockQuadTransform emissiveBlockQuadTransform = new EmissiveBakedModel.EmissiveBlockQuadTransform();
-	public final EmissiveBakedModel.EmissiveItemQuadTransform emissiveItemQuadTransform = new EmissiveBakedModel.EmissiveItemQuadTransform();
+	public final CtmBlockStateModel.CtmQuadTransform ctmQuadTransform = new CtmBlockStateModel.CtmQuadTransform();
 
 	public final ContinuityFeatureStatesImpl featureStates = new ContinuityFeatureStatesImpl();
 	public final MutableMesh mutableMesh = Renderer.get().mutableMesh();
