+++ src/main/java/me/pepperbell/continuity/client/resource/ModelWrappingHandler.java
@@ -1,21 +1,13 @@
 package me.pepperbell.continuity.client.resource;
 
+import me.pepperbell.continuity.client.model.CtmBlockStateModel;
+import net.minecraft.client.render.item.model.ItemModel;
+import net.minecraft.client.render.model.BlockStateModel;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 
-import com.google.common.collect.ImmutableMap;
-
-import me.pepperbell.continuity.client.model.CtmBakedModel;
-import me.pepperbell.continuity.client.model.EmissiveBakedModel;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelLoadingPlugin;
 import net.fabricmc.fabric.api.client.model.loading.v1.ModelModifier;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.client.render.block.BlockModels;
-import net.minecraft.client.render.model.BakedModel;
-import net.minecraft.client.render.model.MissingModel;
-import net.minecraft.client.util.ModelIdentifier;
-import net.minecraft.registry.Registries;
 import net.minecraft.util.Identifier;
 
 public class ModelWrappingHandler {
@@ -24,12 +16,10 @@ public class ModelWrappingHandler {
 
 	private final boolean wrapCtm;
 	private final boolean wrapEmissive;
-	private final ImmutableMap<ModelIdentifier, BlockState> blockStateModelIds;
 
 	private ModelWrappingHandler(boolean wrapCtm, boolean wrapEmissive) {
 		this.wrapCtm = wrapCtm;
 		this.wrapEmissive = wrapEmissive;
-		blockStateModelIds = createBlockStateModelIdMap();
 	}
 
 	@Nullable
@@ -48,62 +38,39 @@ public class ModelWrappingHandler {
 		instance = null;
 	}
 
-	private static ImmutableMap<ModelIdentifier, BlockState> createBlockStateModelIdMap() {
-		ImmutableMap.Builder<ModelIdentifier, BlockState> builder = ImmutableMap.builder();
-		// Match code of BakedModelManager#toStateMap
-		for (Block block : Registries.BLOCK) {
-			Identifier blockId = block.getRegistryEntry().registryKey().getValue();
-			for (BlockState state : block.getStateManager().getStates()) {
-				ModelIdentifier modelId = BlockModels.getModelId(blockId, state);
-				builder.put(modelId, state);
-			}
-		}
-		return builder.build();
-	}
-
-	public BakedModel wrap(BakedModel model, Identifier id) {
-		if (!id.equals(MissingModel.ID)) {
-			if (wrapEmissive) {
-				model = new EmissiveBakedModel(model);
-			}
-		}
+	public ItemModel wrapItem(ItemModel model, Identifier id) {
+//		if (!id.equals(MissingModel.ID)) {
+//			if (wrapEmissive) {
+//				model = new EmissiveBakedModel(model);
+//			}
+//		}
 		return model;
 	}
 
-	public BakedModel wrapBlock(BakedModel model, ModelIdentifier topLevelId) {
+	public BlockStateModel wrapBlock(BlockStateModel model) {
 		if (wrapCtm) {
-			BlockState state = blockStateModelIds.get(topLevelId);
-			if (state != null) {
-				model = new CtmBakedModel(model, state);
-			}
+			model = new CtmBlockStateModel(model);
 		}
-		if (wrapEmissive) {
-			model = new EmissiveBakedModel(model);
-		}
-		return model;
-	}
-
-	public BakedModel ensureWrapped(BakedModel model) {
-		if (wrapEmissive && !(model instanceof EmissiveBakedModel)) {
-			return new EmissiveBakedModel(model);
-		}
+//		if (wrapEmissive) {
+//			model = new EmissiveBakedModel(model);
+//		}
 		return model;
 	}
 
 	@ApiStatus.Internal
 	public static void init() {
 		ModelLoadingPlugin.register(pluginCtx -> {
-			pluginCtx.modifyModelAfterBake().register(ModelModifier.WRAP_LAST_PHASE, (model, ctx) -> {
+			pluginCtx.modifyItemModelAfterBake().register(ModelModifier.WRAP_LAST_PHASE, (model, ctx) -> {
 				ModelWrappingHandler wrappingHandler = getInstance();
 				if (wrappingHandler != null) {
-					return wrappingHandler.wrap(model, ctx.id());
+					return wrappingHandler.wrapItem(model, ctx.itemId());
 				}
 				return model;
 			});
 			pluginCtx.modifyBlockModelAfterBake().register(ModelModifier.WRAP_LAST_PHASE, (model, ctx) -> {
 				ModelWrappingHandler wrappingHandler = getInstance();
 				if (wrappingHandler != null) {
-					return wrappingHandler.wrapBlock(model, ctx.id());
+					return wrappingHandler.wrapBlock(model);
 				}
 				return model;
 			});
