+++ src/main/resources/continuity.mixins.json
@@ -8,7 +8,6 @@
     "BakedModelManagerMixin",
     "BlockModelsMixin",
     "FallingBlockEntityRendererMixin",
-    "LayerRenderStateMixin",
     "LifecycledResourceManagerImplMixin",
     "PistonBlockEntityRendererMixin",
     "RenderLayersMixin",
